+++ app/Http/Controllers/StudentController.php
@@ -5,14 +5,29 @@ namespace App\Http\Controllers;
 use App\Models\Student;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class StudentController extends Controller
 {
-
     public function getStudents(Request $request): JsonResponse
     {
-        return response()->json(Student::with(['instruments' ,'parent'])->get());
+        $relations = [
+            'instruments',
+            'parent',
+            'transactions',
+        ];
+
+        if (auth()->checkTable('users')) {
+            $students = $request->get('withTrashed') ?
+                Student::withTrashed()->with($relations)->get() :
+                Student::with($relations)->get();
+        } else {
+            $students = $request->get('withTrashed') ?
+                $request->user()->students()->withTrashed()->with($relations)->get() :
+                $request->user()->students()->with($relations)->get();
+        }
+        return response()->json($students);
     }
 
     public function getStudent(Request $request, $id): JsonResponse
@@ -29,21 +44,29 @@ class StudentController extends Controller
 
     public function createStudent(Request $request): JsonResponse
     {
-        $request->validate([
-            'first_name' => 'required',
-            'last_name' => 'required',
-            'email' => 'required|email|unique:students,email',
-            'password' => 'required',
-            'parent_id' => 'required|exists:parents,id',
-        ]);
+        if ($request->attributes->get('currentGuard') === 'parent') {
+            $request->request->set('parent_id',$request->user()->id);
+        }
         DB::beginTransaction();
         try {
+            $request->validate([
+
+                'first_name' => 'required',
+                'last_name' => 'required',
+                'email' => 'nullable|email|unique:students,email',
+                'password' => 'required',
+                'parent_id' => 'required|exists:parents,id',
+                'username'=> 'required|unique:students,infos->username'
+            ]);
+            $request->request->set('email',!$request->email?
+                $request->username.'@'.env('APP_DOMAIN'):$request->email);
+
             $student = Student::create($request->all());
             DB::commit();
 
-            return  response()->json([
+            return response()->json([
                 "result" => $student,
-                "message" =>  "Student created successfully",
+                "message" => "Student created successfully",
                 "_t" => "success",
             ]);
         } catch (\Exception $e) {
@@ -52,11 +75,37 @@ class StudentController extends Controller
         }
     }
 
-    public function deleteStudent(Request $request, $id)
+    public function deleteStudent(Request $request): JsonResponse
     {
-        $student = Student::find($id);
-        $student->delete();
-        return response()->json($student);
+
+        DB::beginTransaction();
+        try {
+            $request->validate([
+                'id' => 'required|exists:students,id',
+            ]);
+            $student = Student::find($request->id);
+            if ($student->deleted_at) {
+                DB::commit();
+                return response()->json([
+                    'message' => 'Student already deleted',
+                    '_t' => 'warning',
+                ]);
+            }
+            $student->delete();
+            DB::commit();
+            return response()->json([
+                'message' => 'Student deleted successfully',
+                '_t' => 'success',
+            ]);
+        } catch (\Exception $e) {
+            DB::rollBack();
+            return response()->json([
+                'error' => $e->getMessage(),
+                'message' => 'Student not deleted',
+                '_t' => 'error',
+            ], 500);
+        }
+
     }
 
     // attach instrument to student
